[追加]ファイルを分割
